Starting client rewrite
Navbar is now transparent with only sidebar, logo, profile-link and
cart-container components as search component will have use in the
homefront page. Homepage is gonna consist of two to three 100vh height
pages that contain information about the page and products. Sidebar's
background is now transparent and Promotions component is displayed in
the Homemiddle page.

Navbar is now fixed so it follows the user and the outlook of the
homepage is much cleaner, decorated with product category links, few
products for example and navbar. Color scheme is now more cream colored too.