New function to receive streams from API and lot of small fixes
Removed useless assets from src/assets/ directory and added some new
ones.
Changed README from Bookstore to Chocolatiste which can be the app name
for now.
Added @types/webpack-env for require.context()
Added html-loader for relative references and html parsing
"Name" of the application changed to Chocolatiste across sidebar and
logo
Added new paymentSuccess action and reducer to alertSlice

If you go to Checkout immediately, takeCart(cartItems) function is
invoked to save the cart contents before cleaning them in possible
payment. Also when you go to /checkout page, localStorage is checked for
the user, incase it is null, warning is given to login and forms are
disabled.
Small notice about use of images on the application on the footer.

new fetchPicture function in the Promotion api that lets you request an
image by a name that includes the extension and the api sends you the
file as a stream. Functionality going to be changed, this is an early
test of the function.

Homefront__categories div now takes in three categories which it matches
through a filter()

OrderPayment now immediately creates an order once user submits the
stripe form, whether it succeeds or not.

PromotionCard now uses <img /> instead of <div /> so that images can be
dynamically referenced from the assets.

More BEM across profile-dashboard