fetchCartState renamed to fetchCartItems and route changed to /items
fetchCartState is now named fetchCartItems as it is the primary api call used by the client to fetch cart items,
and the route is now /items. I removed the original fetchCartItems from controllers as redundant,
but I kept it in cart.service.ts and renamed it to fetchItems, since addToCart uses it to fetch the cart items.